requirements.txt
streamlit>=1.55.0
openai>=1.59.2
Pillow>=10.4.0
python-dotenv>=1.0.1
bcrypt>=4.0.0
